+++ pkg/apis/deployment/v2alpha1/arango_member_pod_template.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2022 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2023 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -52,6 +52,7 @@ type ArangoMemberPodTemplate struct {
 	PodSpecChecksum string                `json:"podSpecChecksum,omitempty"`
 	Checksum        string                `json:"checksum,omitempty"`
 
+	// deprecated
 	Endpoint *string `json:"endpoint,omitempty"`
 }