+++ pkg/deployment/agency/state.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2022 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2023 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -165,6 +165,87 @@ func (s State) GetDBServerWithLowestShards() Server {
 		}
 	}
 	return resultServer
+}
+
+type ShardDetails struct {
+	ShardID    string
+	Database   string
+	Collection string
+	Servers    Servers
+}
+
+// GetShardDetailsByID returns the ShardDetails for a given ShardID. If the ShardID is not found, the second return value is false
+func (s State) GetShardDetailsByID(id string) (ShardDetails, bool) {
+	// check first in Plan
+	for dbName, db := range s.Plan.Collections {
+		for colName, col := range db {
+			for sName, servers := range col.Shards {
+				if sName == id {
+					return ShardDetails{
+						ShardID:    sName,
+						Database:   dbName,
+						Collection: colName,
+						Servers:    servers,
+					}, true
+				}
+			}
+		}
+	}
+
+	// check in Current
+	for dbName, db := range s.Current.Collections {
+		for colName, col := range db {
+			for sName, shard := range col {
+				if sName == id {
+					return ShardDetails{
+						ShardID:    sName,
+						Database:   dbName,
+						Collection: colName,
+						Servers:    shard.Servers,
+					}, true
+				}
+			}
+		}
+	}
+
+	return ShardDetails{}, false
+}
+
+type ShardStatus struct {
+	IsSynced bool
+}
+
+func (s State) GetShardsStatus() map[string]bool {
+	q := map[string]bool{}
+
+	for dName, d := range s.Plan.Collections {
+		for cName, c := range d {
+			for sName, servers := range c.Shards {
+				q[sName] = s.IsShardInSync(dName, cName, sName, servers)
+			}
+		}
+	}
+
+	return q
+}
+
+func (s State) IsShardInSync(db, col, shard string, servers Servers) bool {
+	dCurrent, ok := s.Current.Collections[db]
+	if !ok {
+		return false
+	}
+
+	cCurrent, ok := dCurrent[col]
+	if !ok {
+		return false
+	}
+
+	sCurrent, ok := cCurrent[shard]
+	if !ok {
+		return false
+	}
+
+	return sCurrent.Servers.InSync(servers)
 }
 
 // PlanServers returns all servers which are part of the plan