+++ pkg/deployment/pod/utils.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2022 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2023 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -43,11 +43,15 @@ func GenerateMemberEndpoint(services service.Inspector, apiObject meta.Object, s
 func GenerateMemberEndpointFromService(svc *core.Service, apiObject meta.Object, spec api.DeploymentSpec, group api.ServerGroup, member api.MemberStatus) (string, error) {
 	if group.IsArangod() {
 		switch method := spec.CommunicationMethod.Get(); method {
-		case api.DeploymentCommunicationMethodDNS:
+		case api.DeploymentCommunicationMethodDNS, api.DeploymentCommunicationMethodHeadlessDNS:
 			return k8sutil.CreateServiceDNSNameWithDomain(svc, spec.ClusterDomain), nil
 		case api.DeploymentCommunicationMethodIP:
 			if svc.Spec.ClusterIP == "" {
 				return "", errors.Newf("ClusterIP of service %s is empty", svc.GetName())
+			}
+
+			if svc.Spec.ClusterIP == core.ClusterIPNone {
+				return k8sutil.CreateServiceDNSNameWithDomain(svc, spec.ClusterDomain), nil
 			}
 
 			return svc.Spec.ClusterIP, nil