+++ pkg/deployment/reconcile/action.register.generated_test.go
@@ -110,6 +110,16 @@ func Test_Actions(t *testing.T) {
 		})
 	})
 
+	t.Run("CleanMemberService", func(t *testing.T) {
+		ActionsExistence(t, api.ActionTypeCleanMemberService)
+		t.Run("Internal", func(t *testing.T) {
+			require.False(t, api.ActionTypeCleanMemberService.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeCleanMemberService.Optional())
+		})
+	})
+
 	t.Run("CleanOutMember", func(t *testing.T) {
 		ActionsExistence(t, api.ActionTypeCleanOutMember)
 		t.Run("Internal", func(t *testing.T) {
@@ -437,6 +447,16 @@ func Test_Actions(t *testing.T) {
 		})
 		t.Run("Optional", func(t *testing.T) {
 			require.False(t, api.ActionTypeRebalancerGenerate.Optional())
+		})
+	})
+
+	t.Run("RebuildOutSyncedShards", func(t *testing.T) {
+		ActionsExistence(t, api.ActionTypeRebuildOutSyncedShards)
+		t.Run("Internal", func(t *testing.T) {
+			require.False(t, api.ActionTypeRebuildOutSyncedShards.Internal())
+		})
+		t.Run("Optional", func(t *testing.T) {
+			require.False(t, api.ActionTypeRebuildOutSyncedShards.Optional())
 		})
 	})