+++ pkg/deployment/reconcile/plan_builder_normal.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2022 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2023 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -49,6 +49,7 @@ func (r *Reconciler) createNormalPlan(ctx context.Context, apiObject k8sutil.API
 		ApplyIfEmpty(r.createScaleUPMemberPlan).
 		// Check for failed members
 		ApplyIfEmpty(r.createMemberFailedRestoreNormalPlan).
+		ApplyIfEmpty(r.createRebuildOutSyncedPlan).
 		// Check for scale up/down
 		ApplyIfEmpty(r.createScaleMemberPlan).
 		// Update status